Move functions from `gitops/beta/run/cmd.go` to corresponding files.
Move some utility functions, related to devbucket, to `install_bucket_source_and_ks.go`.

Remove unused function `WatchAndSync`.
