+++ pkg/run/install_bucket_source_and_ks.go
@@ -2,20 +2,28 @@ package run
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"os"
+	"path/filepath"
 	"strings"
 	"time"
 
 	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1beta2"
 	"github.com/fluxcd/pkg/apis/meta"
 	sourcev1 "github.com/fluxcd/source-controller/api/v1beta2"
+	"github.com/fsnotify/fsnotify"
+	"github.com/minio/minio-go/v7"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	apimeta "k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
 	"sigs.k8s.io/cli-utils/pkg/object"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -111,6 +119,97 @@ func SetupBucketSourceAndKS(log logger.Logger, kubeClient client.Client, namespa
 	return nil
 }
 
+// FindGitRepoDir finds git repo root directory
+func FindGitRepoDir() (string, error) {
+	gitDir := "."
+
+	for {
+		if _, err := os.Stat(filepath.Join(gitDir, ".git")); err == nil {
+			break
+		}
+
+		gitDir = filepath.Join(gitDir, "..")
+
+		if gitDir == "/" {
+			return "", errors.New("not in a git repo")
+		}
+	}
+
+	return filepath.Abs(gitDir)
+}
+
+// GetRelativePathToRootDir gets relative path to a directory from the git root. It returns an error if there's no git repo.
+func GetRelativePathToRootDir(rootDir string, path string) (string, error) {
+	absGitDir, err := filepath.Abs(rootDir)
+
+	if err != nil { // not in a git repo
+		return "", err
+	}
+
+	return filepath.Rel(absGitDir, path)
+}
+
+// SyncDir recursively uploads all files in a directory to an S3 bucket with minio library
+func SyncDir(log logger.Logger, dir string, bucket string, client *minio.Client) error {
+	log.Actionf("Refreshing bucket %s ...", bucket)
+
+	if err := client.RemoveBucketWithOptions(context.Background(), bucket, minio.RemoveBucketOptions{
+		ForceDelete: true,
+	}); err != nil {
+		// if error is not bucket not found, return error
+		if !strings.Contains(err.Error(), "NoSuchBucket") {
+			return err
+		}
+	}
+
+	if err := client.MakeBucket(context.Background(), bucket, minio.MakeBucketOptions{}); err != nil {
+		return err
+	}
+
+	uploadCount := 0
+	err := filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			log.Failuref("Error walking directory: %v", err)
+			return err
+		}
+
+		if info.IsDir() {
+			// if it's a hidden directory, ignore it
+			if strings.HasPrefix(info.Name(), ".") {
+				return filepath.SkipDir
+			}
+
+			return nil
+		}
+
+		objectName, err := filepath.Rel(dir, path)
+		if err != nil {
+			log.Failuref("Error getting relative path: %v", err)
+			return err
+		}
+		// upload the file
+		_, err = client.FPutObject(context.Background(), bucket, objectName, path, minio.PutObjectOptions{})
+		if err != nil {
+			return err
+		}
+		uploadCount = uploadCount + 1
+		if uploadCount%10 == 0 {
+			fmt.Print(".")
+		}
+		return nil
+	})
+
+	fmt.Println()
+	log.Actionf("Uploaded %d files", uploadCount)
+
+	if err != nil {
+		log.Failuref("Error syncing directory: %v", err)
+		return err
+	}
+
+	return nil
+}
+
 // CleanupBucketSourceAndKS removes the bucket source and ks
 func CleanupBucketSourceAndKS(log logger.Logger, kubeClient client.Client, namespace string) error {
 	const (
@@ -172,8 +271,8 @@ func CleanupBucketSourceAndKS(log logger.Logger, kubeClient client.Client, names
 	return nil
 }
 
-// FindConditionMessages finds the messages in the condition of objects in the inventory.
-func FindConditionMessages(kubeClient client.Client, ks *kustomizev1.Kustomization) ([]string, error) {
+// findConditionMessages finds the messages in the condition of objects in the inventory.
+func findConditionMessages(kubeClient client.Client, ks *kustomizev1.Kustomization) ([]string, error) {
 	if ks.Status.Inventory == nil {
 		return nil, fmt.Errorf("inventory is nil")
 	}
@@ -233,4 +332,131 @@ func FindConditionMessages(kubeClient client.Client, ks *kustomizev1.Kustomizati
 	}
 
 	return messages, nil
+}
+
+func WatchDirsForFileWalker(watcher *fsnotify.Watcher) func(path string, info os.FileInfo, err error) error {
+	return func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return fmt.Errorf("error walking path: %v", err)
+		}
+
+		if info.IsDir() {
+			// if it's a hidden directory, ignore it
+			if strings.HasPrefix(info.Name(), ".") {
+				return filepath.SkipDir
+			}
+
+			if err := watcher.Add(path); err != nil {
+				return err
+			}
+		}
+
+		return nil
+	}
+}
+
+// ReconcileDevBucketSourceAndKS reconciles the dev-bucket and dev-ks asynchronously.
+func ReconcileDevBucketSourceAndKS(log logger.Logger, kubeClient client.Client, namespace string, timeout time.Duration) error {
+	const interval = 3 * time.Second / 2
+
+	// reconcile dev-bucket
+	sourceRequestedAt, err := requestReconciliation(context.Background(), kubeClient,
+		types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-bucket",
+		}, schema.GroupVersionKind{
+			Group:   "source.toolkit.fluxcd.io",
+			Version: "v1beta2",
+			Kind:    "Bucket",
+		})
+	if err != nil {
+		return err
+	}
+
+	// wait for the reconciliation of dev-bucket to be done
+	if err := wait.Poll(interval, timeout, func() (bool, error) {
+		devBucket := &sourcev1.Bucket{}
+		if err := kubeClient.Get(context.Background(), types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-bucket",
+		}, devBucket); err != nil {
+			return false, err
+		}
+
+		return devBucket.Status.GetLastHandledReconcileRequest() == sourceRequestedAt, nil
+	}); err != nil {
+		return err
+	}
+
+	// wait for devBucket to be ready
+	if err := wait.Poll(interval, timeout, func() (bool, error) {
+		devBucket := &sourcev1.Bucket{}
+		if err := kubeClient.Get(context.Background(), types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-bucket",
+		}, devBucket); err != nil {
+			return false, err
+		}
+		return apimeta.IsStatusConditionPresentAndEqual(devBucket.Status.Conditions, meta.ReadyCondition, metav1.ConditionTrue), nil
+	}); err != nil {
+		return err
+	}
+
+	// reconcile dev-ks
+	ksRequestedAt, err := requestReconciliation(context.Background(), kubeClient,
+		types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-ks",
+		}, schema.GroupVersionKind{
+			Group:   "kustomize.toolkit.fluxcd.io",
+			Version: "v1beta2",
+			Kind:    "Kustomization",
+		})
+	if err != nil {
+		return err
+	}
+
+	if err := wait.Poll(interval, timeout, func() (bool, error) {
+		devKs := &kustomizev1.Kustomization{}
+		if err := kubeClient.Get(context.Background(), types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-ks",
+		}, devKs); err != nil {
+			return false, err
+		}
+
+		return devKs.Status.GetLastHandledReconcileRequest() == ksRequestedAt, nil
+	}); err != nil {
+		return err
+	}
+
+	devKs := &kustomizev1.Kustomization{}
+	devKsErr := wait.Poll(interval, timeout, func() (bool, error) {
+		if err := kubeClient.Get(context.Background(), types.NamespacedName{
+			Namespace: namespace,
+			Name:      "dev-ks",
+		}, devKs); err != nil {
+			return false, err
+		}
+
+		healthy := apimeta.IsStatusConditionPresentAndEqual(
+			devKs.Status.Conditions,
+			kustomizev1.HealthyCondition,
+			metav1.ConditionTrue,
+		)
+		return healthy, nil
+	})
+
+	if devKsErr != nil {
+		messages, err := findConditionMessages(kubeClient, devKs)
+		if err != nil {
+			return err
+		}
+
+		for _, msg := range messages {
+			log.Failuref(msg)
+		}
+	}
+
+	return devKsErr
 }
